# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RobloxAI Agent API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tools/analyze-script:
    post:
      operationId: analyzeScript
      tags: [tools]
      summary: Analyze a Lua script for issues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScriptAnalysisInput"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScriptAnalysisResult"
  /tools/remote-scanner:
    post:
      operationId: scanRemotes
      tags: [tools]
      summary: Scan script for RemoteEvents/Functions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoteScanInput"
      responses:
        "200":
          description: Scan result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RemoteScanResult"
  /tools/error-diagnostic:
    post:
      operationId: diagnoseError
      tags: [tools]
      summary: Diagnose Roblox executor error log
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ErrorDiagnosticInput"
      responses:
        "200":
          description: Diagnosis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDiagnosticResult"
  /tools/profiler:
    post:
      operationId: profileScript
      tags: [tools]
      summary: Profile script for performance issues
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfilerInput"
      responses:
        "200":
          description: Profiler result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfilerResult"
  /tools/gui-builder:
    post:
      operationId: buildGui
      tags: [tools]
      summary: Generate GUI JSON config from prompt or image analysis
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuiBuildInput"
      responses:
        "200":
          description: GUI config result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuiBuildResult"
components:
  schemas:
...
    ScriptAnalysisInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
        conversationId:
          type: ["integer", "null"]
    ScriptAnalysisResult:
      type: object
      required: [issues, optimizations, score, fixedCode, explanation]
      properties:
        issues:
          type: array
          items:
            type: string
        optimizations:
          type: array
          items:
            type: string
        score:
          type: integer
        fixedCode:
          type: string
        explanation:
          type: string
    RemoteScanInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    RemoteScanResult:
      type: object
      required: [remotes, suggestions]
      properties:
        remotes:
          type: array
          items:
            $ref: "#/components/schemas/RemoteEntry"
        suggestions:
          type: array
          items:
            type: string
    RemoteEntry:
      type: object
      required: [name, type, path]
      properties:
        name:
          type: string
        type:
          type: string
        path:
          type: string
        args:
          type: array
          items:
            type: string
        bypassSuggestion:
          type: ["string", "null"]
    ErrorDiagnosticInput:
      type: object
      required: [errorLog]
      properties:
        errorLog:
          type: string
        originalCode:
          type: ["string", "null"]
    ErrorDiagnosticResult:
      type: object
      required: [rootCause, fixedCode, explanation, line]
      properties:
        rootCause:
          type: string
        fixedCode:
          type: string
        explanation:
          type: string
        line:
          type: ["integer", "null"]
    ProfilerInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    ProfilerResult:
      type: object
      required: [issues, optimizedCode, fpsImpact, warnings]
      properties:
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ProfilerIssue"
        optimizedCode:
          type: string
        fpsImpact:
          type: string
        warnings:
          type: array
          items:
            type: string
    ProfilerIssue:
      type: object
      required: [type, description, line, severity]
      properties:
        type:
          type: string
        description:
          type: string
        line:
          type: ["integer", "null"]
        severity:
          type: string
    GuiBuildInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
        imageBase64:
          type: ["string", "null"]
        style:
          type: ["string", "null"]
    GuiBuildResult:
      type: object
      required: [guiJson, luaCode, preview]
      properties:
        guiJson:
          type: string
        luaCode:
          type: string
        preview:
          type: string