# lib/api-zod/src/generated/types/errorDiagnosticInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RobloxAI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorDiagnosticInput {
  errorLog: string;
  /** @nullable */
  originalCode?: string | null;
}