# lib/api-zod/src/generated/types/errorDiagnosticResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RobloxAI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorDiagnosticResult {
  rootCause: string;
  fixedCode: string;
  explanation: string;
  /** @nullable */
  line: number | null;
}