# lib/api-zod/src/generated/types/guiBuildInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RobloxAI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface GuiBuildInput {
  prompt: string;
  /** @nullable */
  imageBase64?: string | null;
  /** @nullable */
  style?: string | null;
}