# lib/api-zod/src/generated/types/profilerIssue.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RobloxAI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ProfilerIssue {
  type: string;
  description: string;
  /** @nullable */
  line: number | null;
  severity: string;
}