# lib/api-zod/src/generated/types/remoteEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RobloxAI Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface RemoteEntry {
  name: string;
  type: string;
  path: string;
  args?: string[];
  /** @nullable */
  bypassSuggestion?: string | null;
}