# lib/api-zod/src/generated/types/remoteScanResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RobloxAI Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { RemoteEntry } from './remoteEntry';

export interface RemoteScanResult {
  remotes: RemoteEntry[];
  suggestions: string[];
}